refactor: move asset/storage operations from pipe/components to utils/asset_util
  Separate S3 path building, upload/download, and directory setup functions
  out of pipe/components.py into marie/utils/asset_util.py to eliminate the
  circular import between asset_util and pipe/components.
