--- marie/executor/asset_util.py
@@ -1,192 +1,0 @@
-import functools
-import hashlib
-import os
-import shutil
-import time
-from datetime import datetime
-from pathlib import Path
-from typing import Callable, List
-
-from PIL import Image
-
-from marie.common.file_io import get_cache_dir
-from marie.pipe.components import download_asset
-from marie.utils.image_utils import hash_frames_fast
-from marie.utils.utils import ensure_exists
-
-
-def avoid_concurrent_lock_wrapper(func: Callable) -> Callable:
-    """
-    Wrap the function around a File Lock to ensure that the function is run by a single replica on the same machine.
-
-    :param func: The function to decorate
-    :return: The wrapped function
-    """
-
-    @functools.wraps(func)
-    def arg_wrapper(*args, **kwargs):
-        from marie.importer import ImportExtensions
-
-        with ImportExtensions(
-            required=False,
-            help_text='FileLock is needed to guarantee non-concurrent execution of the function.',
-        ):
-            import filelock
-
-            locks_root = Path.home().joinpath('.locks')
-            locks_root.mkdir(parents=True, exist_ok=True)
-            lock_file = locks_root.joinpath(f'{func.__name__}.lock')
-
-            file_lock = filelock.FileLock(lock_file, timeout=-1)
-
-        with file_lock:
-            try:
-                print(f'Acquired lock for {func.__name__} with {file_lock}')
-                return func(*args, **kwargs)
-            finally:
-                print(f'Released lock for {func.__name__}')
-
-    return arg_wrapper
-
-
-@avoid_concurrent_lock_wrapper
-def prepare_asset_directory(
-    frames, local_path: str, ref_id: str, ref_type: str, logger
-):
-    """
-    Prepares the asset directory by creating the required subdirectories and processing input files.
-
-    :param frames: List of frames (numpy arrays) to process.
-    :param local_path: Local file path of the downloaded S3 file.
-    :param ref_id: Unique identifier for the asset reference.
-    :param ref_type: Type of the reference for the asset.
-    :param logger: Logger instance to handle logging.
-    :return: Tuple containing root asset directory path, frames directory path, and metadata file path.
-    :raises ValueError: If the local_path parameter is None.
-    """
-
-    if local_path is None:
-        logger.error("The 'local_path' parameter is None. Unable to proceed.")
-        raise ValueError("The 'local_path' parameter cannot be None.")
-
-    root_asset_dir = create_working_dir(frames, ref_id=ref_id, ref_type=ref_type)
-    frames_dir = os.path.join(root_asset_dir, "frames")
-    ensure_exists(frames_dir)
-
-    existing_files = (
-        sorted(os.listdir(frames_dir)) if os.path.exists(frames_dir) else []
-    )
-
-    burst_dir = os.path.join(root_asset_dir, "burst")
-    if not existing_files and os.path.exists(burst_dir):
-        burst_files = sorted([f for f in os.listdir(burst_dir) if f.endswith('.tif')])
-        if burst_files:
-            for idx, file in enumerate(burst_files):
-                src = os.path.join(burst_dir, file)
-                dst = os.path.join(frames_dir, f"{idx + 1:05}.png")
-                shutil.copy2(src, dst)
-            existing_files = sorted(os.listdir(frames_dir))
-
-    if existing_files:
-        existing_frames = [
-            os.path.join(frames_dir, file)
-            for file in existing_files
-            if file.endswith('.png')
-        ]
-        valid_frames = len(existing_frames) == len(frames) and all(
-            os.path.isfile(os.path.join(frames_dir, f"{idx + 1:05}.png"))
-            for idx in range(len(frames))
-        )
-
-        if valid_frames:
-            logger.info(
-                f"Frames already exist in '{frames_dir}' and match the expected format. Skipping further processing."
-            )
-            metadata_file = os.path.join(root_asset_dir, f"{ref_id}.meta.json")
-            return root_asset_dir, frames_dir, metadata_file
-
-    # Copy local file to the target path in the asset directory
-    target_path = os.path.join(root_asset_dir, ref_id)
-    if not os.path.exists(target_path):
-        shutil.copy2(local_path, target_path)
-        with open(target_path, 'a') as f:
-            f.flush()
-            os.fsync(f.fileno())
-        logger.info(f"Copied file from '{local_path}' to '{target_path}'.")
-
-    logger.info(f"Root asset directory created: '{root_asset_dir}'")
-
-    for idx, frame in enumerate(frames):
-        frame_path = os.path.join(frames_dir, f"{idx + 1:05}.png")
-        try:
-            Image.fromarray(frame).save(frame_path)
-            logger.debug(f"Frame {idx + 1} saved at '{frame_path}'.")
-
-            img = Image.open(frame_path)
-            logger.debug(f"Image dimensions: {img.size}")
-        except Exception as e:
-            logger.error(f"Error while processing frame {idx + 1} - {e}")
-            raise
-
-    # Download additional metadata for the asset
-    metadata_file = download_asset(
-        ref_id=ref_id,
-        ref_type=ref_type,
-        root_asset_dir=root_asset_dir,
-        s3_file_path=f"{ref_id}.meta.json",
-        overwrite=True,
-    )
-    logger.info(f"Metadata file downloaded and stored at: '{metadata_file}'")
-    time.sleep(0.1)  # Ensure file system operations are completed
-
-    # Ensure the metadata file exists and that it is a valid JSON file
-    if not os.path.exists(metadata_file):
-        logger.error(f"Metadata file '{metadata_file}' does not exist.")
-        raise FileNotFoundError(f"Metadata file '{metadata_file}' not found.")
-
-    try:
-        with open(metadata_file, 'r') as f:
-            metadata = f.read()
-            if not metadata.strip().startswith('{'):
-                raise ValueError(
-                    f"Metadata file '{metadata_file}' is not a valid JSON."
-                )
-    except Exception as e:
-        logger.error(f"Error reading metadata file '{metadata_file}': {e}")
-        raise
-
-    return root_asset_dir, frames_dir, metadata_file
-
-
-def create_working_dir(
-    frames: List,
-    backup: bool = False,
-    ref_id: str = None,
-    ref_type: str = None,
-    queue_id: str = None,
-    job_id: str = None,
-) -> str:
-    """Create an isolated working directory for a pipeline execution.
-
-    The directory hash is derived from the frame pixel data combined with
-    request identifiers (ref_id, ref_type, queue_id, job_id) so that
-    concurrent jobs never collide even when processing identical frames.
-    """
-    frame_checksum = hash_frames_fast(frames=frames)
-    md5 = hashlib.md5(frame_checksum.encode("utf-8"))
-
-    for token in (ref_id, ref_type, queue_id, job_id):
-        if token:
-            md5.update(token.encode("utf-8"))
-    combined_checksum = md5.hexdigest()
-
-    generators_dir = os.path.join(get_cache_dir(), "generators")
-    os.makedirs(generators_dir, exist_ok=True)
-    target_dir = os.path.join(generators_dir, combined_checksum)
-
-    # create backup name by appending a timestamp
-    if backup and os.path.exists(target_dir):
-        ts = datetime.now().strftime("%Y%m%d%H%M%S")
-        shutil.move(target_dir, f"{target_dir}-{ts}")
-
-    return ensure_exists(target_dir)
