+++ marie/executor/extract/document_annotator_executor.py
@@ -10,7 +10,6 @@ from omegaconf import OmegaConf
 from marie.api.docs import AssetKeyDoc
 from marie.constants import __config_dir__
 from marie.excepts import RuntimeTerminated
-from marie.executor.asset_util import prepare_asset_directory
 from marie.executor.extract.util import layout_config
 from marie.executor.marie_executor import MarieExecutor
 from marie.executor.mixin import StorageMixin
@@ -21,6 +20,7 @@ from marie.logging_core.logger import MarieLogger
 from marie.logging_core.mdc import MDC
 from marie.logging_core.predefined import default_logger as logger
 from marie.models.utils import torch_gc
+from marie.utils.asset_util import prepare_asset_directory
 from marie.utils.docs import docs_from_asset, frames_from_docs
 from marie.utils.json import load_json_file
 from marie.utils.network import get_ip_address
