+++ marie/executor/extract/document_annotator_table_parser_executor.py
@@ -7,9 +7,6 @@ from omegaconf import OmegaConf
 
 from marie import requests
 from marie.api.docs import AssetKeyDoc
-from marie.executor.asset_util import (
-    prepare_asset_directory,
-)
 from marie.executor.extract import DocumentAnnotatorExecutor
 from marie.executor.extract.util import layout_config, setup_table_directories
 from marie.extract.readers.meta_reader.meta_reader import MetaReader
@@ -18,6 +15,9 @@ from marie.extract.results.result_parser import extract_tables, highlight_tables
 from marie.extract.structures import UnstructuredDocument
 from marie.logging_core.logger import MarieLogger
 from marie.logging_core.predefined import default_logger as logger
+from marie.utils.asset_util import (
+    prepare_asset_directory,
+)
 from marie.utils.docs import docs_from_asset, frames_from_docs
 from marie.utils.json import load_json_file
