+++ marie/executor/pipeline/document_llm_pipeline_executor.py
@@ -5,7 +5,6 @@ from docarray import DocList
 from marie import requests, safely_encoded
 from marie.api import AssetKeyDoc, value_from_payload_or_args
 from marie.boxes import PSMode
-from marie.executor.asset_util import create_working_dir
 from marie.executor.pipeline.document_pipeline_executor import PipelineExecutor
 from marie.executor.request_util import get_frames_from_docs, parse_parameters
 from marie.logging_core.mdc import MDC
@@ -13,6 +12,7 @@ from marie.logging_core.predefined import default_logger as logger
 from marie.models.utils import torch_gc
 from marie.ocr import CoordinateFormat
 from marie.pipe.llm_pipeline import LLMPipeline
+from marie.utils.asset_util import create_working_dir
 
 
 # TODO: refactor and aggregate into functional packages
